General rule of thumb; never commit dependencies
There might be exceptions for other types of projects but when it comes
to node and node_modules, just shouldn't commit the entire dependency
tree. This makes the repository enormous and dependencies near
impossible to manage. Notice that the directory is in the .gitignore
file and should be ignored.

This is what the package-lock.json file is for. It keeps a record of
which versions of your dependencies you are using so when I or a build
server runs `npm install` we all have the same versions. This file you
should definitely commit.

Since the node_modules/ folder is already in the git history you should
run `git filter-branch --index-filter 'git rm -r --cached
--ignore-unmatch food-delivery/node_modules/' -- --all` or something
similar to purge it.